chore: reorganize admin pages
