+++ src/pages/admin/employees/EmployeesPage.tsx
@@ -5,13 +5,13 @@ import { Button } from "@/components/ui/button";
 import { Plus, Upload, Users, UserCheck, UserX, Filter } from "lucide-react";
 import { useNavigate } from "react-router-dom";
 import { useEmployees } from "@/hooks/useEmployees";
-import { EmployeeFilters } from "./EmployeeFilters";
+import { EmployeeFilters } from "../../../components/admin/employees/EmployeeFilters";
 import { PageHeader } from "@/components/ui/page-header";
 import { StatCard } from "@/components/ui/stat-card";
 import { FilterSection } from "@/components/ui/filter-section";
 import { useEmployeeStore } from "@/stores/employeeStore";
 import { Skeleton } from "@/components/ui/skeleton";
-import { EmployeeTableMemo } from "./EmployeeTableMemo";
+import { EmployeeTableMemo } from "../../../components/admin/employees/EmployeeTableMemo";
 import { MobileTable, MobileTableRow } from "@/components/ui/mobile-table";
 import { Badge } from "@/components/ui/badge";
 import { useMobileResponsive } from "@/hooks/use-mobile-responsive";
